mode/hint: Allow acting on multiple inputs/buttons etc.
Before this change, selecting several inputs in follow-hint was only activating
the first one. This change makes all the specialized follow hint methods to
invoke %follow-hint on any non-link element. This way, all the inputs/buttons
are activated, while all the links are opened.

No downsides to this change, it's an improvement over existing behavior.