Support spaces in EXTERNAL-EDITOR-PROGRAM
`VISUAL` and `EDITOR` might contain parameters, and we cannot mix and
match strings and lists in `uiop:run-program` as it signals an error. We
need explicit code paths for each.

Example of new behavior when adding the argument "path":

| editor                | command                      |
|-----------------------|------------------------------|
| "editor --param"      | "editor --param path"        |
| '("editor" "--param") | '("editor" "--param" "path") |

The old behavior was

| editor                | command                      |
|-----------------------|------------------------------|
| "editor --param"      | '("editor --param" "path")   |
| '("editor" "--param") | '("editor" "--param" "path") |

`'("editor --param" "path")'` signals an error as it cannot find the
command "editor --param".