prompt-buffer: Remove inactive buttons logic.
The previous commit introduces a UI glitch where the left arrow down button is
drawn as inactive by a brief moment.  The root cause isn't related to the
implementation of the feature itself, but due to the poor architecture design of
update-prompt-input.

Steps to reproduce:

- Invoke execute-command
- Run `(update-prompt-input (current-prompt-buffer) "fo")`

The issue can be understood by inspecting the calls to
`(prompt-render-suggestions prompt-buffer)` in `update-prompt-input`.