.github/workflows/tests: Major refactor.
Deprecate "Keymap type checking" and "Validate make-instance symbols".

Deprecate "Build Nyxt executable" since, in its current implementation, it's not
clear whether it is produced.  The generated binary can't be tried out locally
since it is generated in an Ubuntu environment.  Answers to questions such as
"did the tests pass?" or "did it compile?" are more meaningful than "was the
binary created?".

Deprecate caching roswell since that job succeeds in a matter of seconds.  The
job that could benefit from caching is git submodule fetching.