parenscript-macro: Optimize 'rqsa' and sort elements depth first
Unlike the macro 'rqs-nyxt-id', we don't make use of 'nyxt-shadow-root'
attribute because the macro 'rqsa' might get called befoore the DOM is update
through the function 'update-document-model'.
We use the TreeWalker for this algorithm based on the benchmarks of the
following StackOverflow response https://stackoverflow.com/a/64551276

Ensure that the results produced by rqsa are sorted in depth first algorithm so
that it matches the generated hints order which is produced in the same order
according to the doocumentation of the function 'clss:select'.

Note that sorting the output array based on 'nyxt-identifier' is not possible
because elements that are dynamically added to the DOM are assigned a
'nyxt-identifier' that is greater than the last generated identifier, so it
would break the depth-first order. Add to that the fact that 'rqsa' macro might
be called on elements that do not have nyxt-attributes as mentioned previously.