input: Do not block on input dispatch.
Otherwise, it wouldn't be possible to execute JavaScript when handling input
from the renderer (Electron port).

In more detail:

- JavaScript waits for a synchronous response from Nyxt as to whether it should
prevent an event from bubbling, or consume it. In other words, JavaScript needs
to know whether it needs to run event.preventDefault().

- In our input handling on the Lisp side, if we try to execute some JavaScript
before we return t/nil we'll send a request to a blocked JavaScript program (the
one waiting for synchronous input), asking it to do something.

- Our Lisp program will not continue until the JavaScript we tried to execute
runs.

- Our Lisp program will be unable to return t/nil.

- We have deadlocked our program.