Disable custom fonts.
Both on Electron+macOS and Electron+Linux custom fonts aren't working.

On Electron+Linux, no errors are raised, whereas on Electron+macOS the opposite
is observed.  It's odd, but we'll fix custom fonts on the Electron port anyway
on both settings in the future.

The changes have no impact on the webkitGTK port.