Deprecate use of nclasses:define-generic.
The macro obscures the codebase for no good reason.  It was introduced to "fix"
the fact that when defining a method with a docstring, the automatically
generated lacks the documentation.  While it may be a (small) issue when a
single method is defined without an explicit defgeneric, it should prompt us to
properly use the well-established and specified CL constructs.

define-parenscript{,-async} are defined as methods.  The motivation is to ease
running arbitrary code around JS routines.  However, it is being leveraged in a
single instance (focus-select-element) to set vi-modes.

Closes #3422.