Refactor fullscreen and maximized window logic.
The new foreign interface architecture handles the state of the relevant window
slots (fullscreen-p, maximized-p) instead of setf-ing them on the renderer
packages.  This replaces the need for the :skip-renderer-resize keyword.

Fixed bug in toggle-message-buffer, where the height of the buffer was being
probed incorrectly.

Refactor the handling of the window-state-event in GTK.  It didn't account for
events where the window is no longer fullscreen or maximized.

Simplify implementation of toggle-fullscreen, toggle-maximize,
toggle-status-buffer and toggle-message-buffer.

As for the fullscreen rationale regarding the UI: when a fullscreen event is
emitted by JS's Fullscreen API, the status and message buffers are hidden.  For
all other fullscreen events, the status and message buffers are kept.  Note that
we can't hide the message buffer on fullscreen since then there would be no way
to get that important piece of information.  On typical browsers, the equivalent
of the message buffer is a dynamically visible toolbar at the bottom left (and
it is still displayed on fullscreen).