Refactor fullscreen and maximized window logic.
The new foreign interface architecture handles the state of the relevant window
slots (such as fullscreen-p) instead of doing that on the renderer packages.
This replaces the need for the :skip-renderer-resize keyword.

Fixed bug in toggle-message-buffer, where the height of the buffer was being
probed incorrectly.

Refactor how the window-state-event in handled in GTK.

Simplify implementation of toggle-fullscreen, toggle-maximize,
toggle-status-buffer and toggle-message-buffer.

In terms of the UI:

- a fullscreen invoked by the user or by the window manager expands the UI
fully (keeping the status and message buffers);

- a fullscreen requested by the renderer (e.g. fullscreen video stream) hide the
status and message buffers.