start: Fix socket permissions.
It's important to set the directory containing the sockets (see below).

The permission of the sockets themselves are set as to lean on the side of
caution, when it comes to security.

From (man "unix"):

In the Linux implementation, pathname sockets honor the permissions of the
directory they are in.  Creation of a new socket fails if the process does not
have write and search (execute) permission on the directory in which the socket
is created.

On Linux, connecting to a stream socket object requires write permission on that
socket; sending a datagram to a datagram socket likewise requires write
permission on that socket.  POSIX does not make any statement about the effect
of the permissions on a socket file, and on some systems (e.g., older BSDs), the
socket permissions are ignored.  Portable programs should not rely on this
feature for security.