Re-implement browsing history logic and its persistence.
Add ffi-buffer-navigate-*, as to rely on the renderer to traverse history.

The global history tree has shown us that it does not work, and is not worth
it. This simplification is the first step in building something reliable.

As soon as Nyxt is stable, we will resume with the per-buffer history tree. So
we will bring back the old behavior before the global history tree.