Refactor on-signal-load-finished signature.
Fix a bug where the history file is populated by entries whose URL and title
conflict with each other.

The only problem that can occur now is due to the architecture of our system. We
can have the following situation; the user has a lot of history, they press
"navigate-backwards" very very quickly. The on-signal-finished event will be
fired by electron, but we may not query electron quickly enough for the URL and
Title. That is because we are handling the response in Lisp. Unless we update
cl-electron to add a custom handler to did-finish-load, I don't think we could
get this data synchronously. This is a very minor bug, and it can be fixed
later. I'll add a note to the cl-electron repository.

See atlas-engineer/cl-electron#60.