prompt-buffer: Limit prompt length, set minimum width for prompt input.
When we are dynamically setting the prompt input to include information, we
cannot have it take an infinite amount of space, at some point we should cut off
the text. This will be particularly useful for querying the user for information
about specific URLs, which can have a very long length.