Delete sly/slime dependencies.
It makes little sense to depend on them for the purpose of instrospectability.
A smaller library (with a limited scope) would be wiser.

The functionality of commands start-swank and start-slynk should be maintained
by a Nyxt extension.  Obviously, Nyxt can still be started from a REPL when the
source is available.