prompt-buffer: Fix bug involving clicking on suggestions.
When clicking on suggestions outside the currently active source, the system
would fail. This is because merely the index of a suggestion was considered
rather than the source AND the index together.

Additionally: Disable disabling mouse support. This is not supported anywhere
else in the interface and belongs to a third party plugin if anything.