Delete reduce-tracking mode
Difficult to implement in the current cl-electron renderer. Additionally, it is
debatable whether this method is actually effective. Finally, it is not one of
the stated design goals of Nyxt to be ultra private. There exist other projects
for that purpose. Nyxt is first and foremost a hackable browser that you can
mold to whatever you are trying to do.