Move GTK specific modes to GTK system
At the current time, these modes are not yet implemented in Electron. They may
be implemented in the future. Additionally, for many of them their behavior is
session wide instead of view specific. This would mean potentially proliferating
several views for each buffer. This is possible, but with debatable merit.

Additionally, they are removed from the manual/tutorial, which would otherwise
not compile.